# Lab5/pzpi-22-8-khudanova-polina-lab5/openapi.yaml
openapi: 3.0.0
info:
  title: Pet Health API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /auth/register:
    post:
      summary: Реєстрація користувача
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успішна реєстрація
  /auth/login:
    post:
      summary: Логін користувача
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успішний вхід
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /pets:
    get:
      summary: Отримати список тварин
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список тварин
    post:
      summary: Додати нову тварину
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, species, breed]
              properties:
                name:
                  type: string
                species:
                  type: string
                breed:
                  type: string
      responses:
        '200':
          description: Тварина додана
  /pets/{id}:
    put:
      summary: Оновити тварину
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                species:
                  type: string
                breed:
                  type: string
      responses:
        '200':
          description: Тварина оновлена
    delete:
      summary: Видалити тварину
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Тварина видалена

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT